requirements.txt
# native dependencies
Pillow==7.2.0
lxml==4.5.2
defusedxml==0.6.0
psycopg2==2.8.5
Django==2.2.15

# Other
amqp==2.6.1
pyyaml>=4.2b1
beautifulsoup4==4.9.1
httplib2<0.18.2
hyperlink==20.0.1
idna>=2.5,<2.11
urllib3==1.25.10
Paver==1.3.4
python-slugify==4.0.1
decorator==4.4.2
celery==4.4.7
kombu==4.6.11
boto3==1.14.39
six==1.15.0
tqdm==4.48.2
Deprecated==1.2.10
wrapt==1.12.1

# Django Apps
django-allauth==0.42.0
django-appconf==1.0.4
django-filter==2.3.0
django-imagekit==4.0.2
django-jsonfield==1.4.0
django-jsonfield-compat==0.4.4
django-taggit==1.3.0
django-markdownify==0.8.1
django-mptt==0.11.0
django-modeltranslation>=0.11,<0.16.0
django-treebeard==4.3.1
django-guardian<2.3.1
django-downloadview==1.10
django-polymorphic==2.1.2
django-tastypie<0.15.0
django-tinymce==3.0.2
django-grappelli==2.14.2
oauthlib==3.1.0
pyopenssl==19.1.0

# geopython dependencies
pyproj==2.6.1.post1
OWSLib==0.20.0
pycsw==2.4.2
SQLAlchemy==1.3.18 # required by PyCSW
Shapely==1.7.0
mercantile==1.1.5
geoip2==3.0.0
numpy==1.19.*

# # Apps with packages provided in GeoNode's PPA on Launchpad.

# Django Apps
dj-database-url==0.5.0
dj-pagination==2.5.0
pinax-notifications==6.0.0
pinax-ratings==4.0.0
django-celery-monitor<=1.1.2
django-celery-beat==2.0.0
django-floppyforms<1.10.0
django-forms-bootstrap<=3.1.0
geonode-django-activity-stream==0.9.0
django-autocomplete-light==3.5.1
django-leaflet==0.27.1
django-invitations<1.9.4
django-recaptcha==2.0.6
geonode-oauth-toolkit==2.0.0

# GeoNode org maintained apps.
django_mapstore_adapter==2.0.4.6
django-geonode-mapstore-client==2.0.8.1
django-geonode-client==1.0.9
geonode-user-messages==2.0.1
geonode-avatar==5.0.7
geonode-announcements==2.0.2
gn-arcrest==10.5.1
geonode-dialogos==3.0.1
geoserver-restconfig==2.0.4.6
gn-gsimporter==2.0.1
gisdata==0.5.4

# haystack/elasticsearch
django-haystack==2.8.1
elasticsearch>=2.0.0,<8.0.0

# datetimepicker widget
django-bootstrap3-datetimepicker-2==2.8.2

# AWS S3 dependencies
django-storages==1.9.1

# Django Caches
python-memcached<=1.59

# Contribs
xmltodict<0.12.1
inflection<0.5.1
jdcal==1.4.1
mock<5.0.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
timeout-decorator==0.4.1
python_resize_image==1.1.19

# required by monitoring
psutil==5.7.2
django-cors-headers==3.4.0
user-agents
xmljson
django-ipware<3.1
# no version here, use latest one with fresh data
pycountry

# production
uWSGI==2.0.19.1
gunicorn==20.0.4
ipython
docker==4.3.0
invoke==1.4.1

# tests
coverage==5.2.1
parse-type==0.5.2
requests-toolbelt==0.9.1
flake8==3.8.3
pytest==6.0.1
pytest-bdd==3.4.0
splinter==0.13.0
pytest-splinter==2.1.0
pytest-django==3.9.0
setuptools==49.2.0
pip==20.1.1
Twisted==20.3.0
factory-boy==2.12.0
